Calibration modified to hook up to new Calibration GUI. Added functionality for importing background images. Made images scalable with dimension/length button. Added grid to map. Added display of lines and other things on mouse movement. Changed redraw to draw to bitmap, then to canvas, which reduces flashing, especially for redrawing Vicon.
