+++ src/RegionEditorFiles/NewRegionEditor/RegionEditorV3.wxg
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<!-- generated by wxGlade 0.6.5 (standalone edition) on Wed Apr 25 12:33:14 2012 -->
+<!-- generated by wxGlade 0.6.5 (standalone edition) on Tue May 01 14:50:10 2012 -->
 
 <application path="C:\Users\cws68\ltlmop\src\RegionEditorFiles\NewRegionEditor\regionEditor.py" name="" class="" option="0" language="python" top_window="regEd" encoding="ISO-8859-15" use_gettext="0" overwrite="0" use_new_namespace="1" for_version="2.8" is_template="0" indent_amount="4" indent_symbol="space" source_extension=".cpp" header_extension=".h">
     <object class="regionEditor" name="regEd" base="EditFrame">
@@ -29,6 +29,11 @@
                         <label>Save As</label>
                         <name>menuSaveAs</name>
                         <handler>OnMenuSaveAs</handler>
+                    </item>
+                    <item>
+                        <label>---</label>
+                        <id>---</id>
+                        <name>---</name>
                     </item>
                     <item>
                         <label>Exit</label>
@@ -65,6 +70,11 @@
                         <handler>OnMenuPoly</handler>
                     </item>
                     <item>
+                        <label>---</label>
+                        <id>---</id>
+                        <name>---</name>
+                    </item>
+                    <item>
                         <label>Add Point</label>
                         <name>menuAddPoint</name>
                         <handler>OnMenuAddPoint</handler>
@@ -75,9 +85,9 @@
                         <handler>OnMenuRemPoint</handler>
                     </item>
                 </menu>
-                <menu name="menuView" label="View">
+                <menu name="menuView" label="Tools">
                     <item>
-                        <label>Markers</label>
+                        <label>Stream Markers</label>
                         <name>menuMarkers</name>
                         <checkable>1</checkable>
                         <handler>OnMenuMarkers</handler>
@@ -88,9 +98,9 @@
                         <handler>OnMenuMarkersClear</handler>
                     </item>
                     <item>
-                        <label>Camera Capture</label>
-                        <name>menuCamera</name>
-                        <handler>OnMenuCamera</handler>
+                        <label>---</label>
+                        <id>---</id>
+                        <name>---</name>
                     </item>
                     <item>
                         <label>Import Background Image</label>
@@ -101,6 +111,25 @@
                         <label>Clear Background Image</label>
                         <name>menuImageClear</name>
                         <handler>OnMenuImageClear</handler>
+                    </item>
+                    <item>
+                        <label>---</label>
+                        <id>---</id>
+                        <name>---</name>
+                    </item>
+                    <item>
+                        <label>Audio Feedback</label>
+                        <name>menuFeedback</name>
+                        <checkable>1</checkable>
+                    </item>
+                    <item>
+                        <label>---</label>
+                        <id>---</id>
+                        <name>---</name>
+                    </item>
+                    <item>
+                        <label>Calibrate</label>
+                        <name>menuCalibrate</name>
                     </item>
                 </menu>
                 <menu name="helpmenu" label="Help">
@@ -131,7 +160,7 @@
                             <border>0</border>
                             <option>0</option>
                             <object class="wxToggleButton" name="toggleVicon" base="EditToggleButton">
-                                <label>Vicon</label>
+                                <label>Markers</label>
                                 <events>
                                     <handler event="EVT_TOGGLEBUTTON">OnToggleVicon</handler>
                                 </events>
@@ -142,10 +171,10 @@
                             <flag>wxALIGN_CENTER_HORIZONTAL|wxALIGN_CENTER_VERTICAL</flag>
                             <border>0</border>
                             <option>1</option>
-                            <object class="wxButton" name="buttonCamera" base="EditButton">
-                                <label>Camera</label>
+                            <object class="wxButton" name="buttonImage" base="EditButton">
+                                <label>Image</label>
                                 <events>
-                                    <handler event="EVT_BUTTON">OnButtonCamera</handler>
+                                    <handler event="EVT_BUTTON">OnButtonImage</handler>
                                 </events>
                                 <size>50, 50</size>
                             </object>
